Expanded charger validation test suite.
